requirements.txt
transformers
unidecode
numpy
pandas
torch
google-generativeai
statsmodels
matplotlib
accelerate
datasets